refactor: split UnitTest1.cs into 6 test files by component
- ChunkSplitterTests.cs (6 tests)
- SymbolExtractorTests.cs (16 tests)
- FileIndexerTests.cs (8 tests)
- DatabaseTests.cs (13 tests)
- DbReaderTests.cs (12 tests)
- McpServerTests.cs (24 tests)

Update CLAUDE.md and DEVELOPER_GUIDE.md architecture sections (EN/JP).
