+++ tests/CodeIndex.Tests/DatabaseTests.cs
@@ -1,0 +1,353 @@
+using CodeIndex.Database;
+using CodeIndex.Models;
+
+namespace CodeIndex.Tests;
+
+/// <summary>
+/// Tests for DbContext and DbWriter integration.
+/// DbContextとDbWriterの統合テスト。
+/// </summary>
+public class DatabaseTests : IDisposable
+{
+    private readonly string _dbPath;
+    private readonly DbContext _db;
+    private readonly DbWriter _writer;
+
+    public DatabaseTests()
+    {
+        _dbPath = Path.Combine(Path.GetTempPath(), $"codeindex_test_{Guid.NewGuid():N}.db");
+        _db = new DbContext(_dbPath);
+        _db.InitializeSchema();
+        _writer = new DbWriter(_db.Connection);
+    }
+
+    [Fact]
+    public void InitializeSchema_CreatesAllTables()
+    {
+        // Verify tables exist by querying sqlite_master
+        // sqlite_masterを問い合わせてテーブルの存在を確認
+        using var cmd = _db.Connection.CreateCommand();
+        cmd.CommandText = "SELECT name FROM sqlite_master WHERE type='table' ORDER BY name";
+        var tables = new List<string>();
+        using var reader = cmd.ExecuteReader();
+        while (reader.Read())
+            tables.Add(reader.GetString(0));
+
+        Assert.Contains("files", tables);
+        Assert.Contains("chunks", tables);
+        Assert.Contains("symbols", tables);
+        Assert.Contains("fts_chunks", tables);
+    }
+
+    [Fact]
+    public void UpsertFile_InsertsAndReturnsId()
+    {
+        var file = new FileRecord
+        {
+            Path = "src/main.py",
+            Lang = "python",
+            Size = 100,
+            Lines = 10,
+            Checksum = "abc123",
+            Modified = new DateTime(2025, 1, 1, 0, 0, 0, DateTimeKind.Utc),
+        };
+
+        var id = _writer.UpsertFile(file);
+        Assert.True(id > 0);
+    }
+
+    [Fact]
+    public void UpsertFile_ReplacesOnConflict()
+    {
+        // Same path should replace (not duplicate)
+        // 同一パスは置換される（重複しない）
+        var file1 = new FileRecord
+        {
+            Path = "src/app.py", Lang = "python", Size = 100, Lines = 10,
+            Modified = new DateTime(2025, 1, 1, 0, 0, 0, DateTimeKind.Utc),
+        };
+        var file2 = new FileRecord
+        {
+            Path = "src/app.py", Lang = "python", Size = 200, Lines = 20,
+            Modified = new DateTime(2025, 2, 1, 0, 0, 0, DateTimeKind.Utc),
+        };
+
+        _writer.UpsertFile(file1);
+        _writer.UpsertFile(file2);
+
+        var (count, _, _) = _writer.GetCounts();
+        Assert.Equal(1, count);
+    }
+
+    [Fact]
+    public void GetUnchangedFileId_ReturnIdIfUnchanged()
+    {
+        var modified = new DateTime(2025, 1, 1, 0, 0, 0, DateTimeKind.Utc);
+        var file = new FileRecord
+        {
+            Path = "src/lib.py", Lang = "python", Size = 50, Lines = 5,
+            Modified = modified,
+        };
+        _writer.UpsertFile(file);
+
+        // Same modified time should return the ID
+        // 同一更新日時ならIDを返す
+        var id = _writer.GetUnchangedFileId("src/lib.py", modified);
+        Assert.NotNull(id);
+
+        // Different modified time should return null
+        // 異なる更新日時ならnullを返す
+        var id2 = _writer.GetUnchangedFileId("src/lib.py", modified.AddHours(1));
+        Assert.Null(id2);
+    }
+
+    [Fact]
+    public void GetUnchangedFileId_MatchesByChecksumWhenTimestampDiffers()
+    {
+        var modified = new DateTime(2025, 1, 1, 0, 0, 0, DateTimeKind.Utc);
+        var checksum = "abc123def456";
+        var file = new FileRecord
+        {
+            Path = "src/checksum.py", Lang = "python", Size = 50, Lines = 5,
+            Modified = modified, Checksum = checksum,
+        };
+        _writer.UpsertFile(file);
+
+        // Different timestamp but same checksum should return the ID (e.g. git checkout)
+        // タイムスタンプ異なるがチェックサム一致ならIDを返す（例: git checkout）
+        var newModified = modified.AddHours(1);
+        var id = _writer.GetUnchangedFileId("src/checksum.py", newModified, checksum);
+        Assert.NotNull(id);
+
+        // Different timestamp AND different checksum should return null
+        // タイムスタンプもチェックサムも異なるならnullを返す
+        var id2 = _writer.GetUnchangedFileId("src/checksum.py", newModified.AddHours(1), "different_checksum");
+        Assert.Null(id2);
+    }
+
+    [Fact]
+    public void InsertChunks_InsertsAndPopulatesFts()
+    {
+        var fileId = _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/test.py", Lang = "python", Size = 100, Lines = 10,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+        });
+
+        var chunks = new List<ChunkRecord>
+        {
+            new() { FileId = fileId, ChunkIndex = 0, StartLine = 1, EndLine = 10, Content = "def authenticate(user):" },
+        };
+        _writer.InsertChunks(chunks);
+
+        // Verify FTS search works / FTS検索が動作することを確認
+        using var cmd = _db.Connection.CreateCommand();
+        cmd.CommandText = "SELECT content FROM fts_chunks WHERE fts_chunks MATCH 'authenticate'";
+        var result = cmd.ExecuteScalar() as string;
+        Assert.NotNull(result);
+        Assert.Contains("authenticate", result);
+    }
+
+    [Fact]
+    public void InsertSymbols_InsertsCorrectly()
+    {
+        var fileId = _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/svc.py", Lang = "python", Size = 50, Lines = 5,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+        });
+
+        var symbols = new List<SymbolRecord>
+        {
+            new() { FileId = fileId, Kind = "function", Name = "process", Line = 1 },
+            new() { FileId = fileId, Kind = "class", Name = "Service", Line = 5 },
+        };
+        _writer.InsertSymbols(symbols);
+
+        var (_, _, symbolCount) = _writer.GetCounts();
+        Assert.Equal(2, symbolCount);
+    }
+
+    [Fact]
+    public void DeleteFileData_RemovesChunksAndSymbols()
+    {
+        var fileId = _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/del.py", Lang = "python", Size = 50, Lines = 5,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+        });
+
+        _writer.InsertChunks([new() { FileId = fileId, ChunkIndex = 0, StartLine = 1, EndLine = 5, Content = "test" }]);
+        _writer.InsertSymbols([new() { FileId = fileId, Kind = "function", Name = "test", Line = 1 }]);
+
+        _writer.DeleteFileData(fileId);
+
+        var (_, chunkCount, symbolCount) = _writer.GetCounts();
+        Assert.Equal(0, chunkCount);
+        Assert.Equal(0, symbolCount);
+    }
+
+    [Fact]
+    public void CleanExistingFileData_PreventsFtsOrphans()
+    {
+        // Insert a file with chunks (populates FTS) / ファイルとチャンク（FTS含む）を挿入
+        var fileId = _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/orphan.py", Lang = "python", Size = 50, Lines = 5,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+        });
+        _writer.InsertChunks([new() { FileId = fileId, ChunkIndex = 0, StartLine = 1, EndLine = 5, Content = "def hello_orphan_test(): pass" }]);
+
+        // Verify FTS has the entry / FTSにエントリがあることを確認
+        using var cmd1 = _db.Connection.CreateCommand();
+        cmd1.CommandText = "SELECT COUNT(*) FROM fts_chunks WHERE fts_chunks MATCH 'hello_orphan_test'";
+        Assert.Equal(1L, (long)cmd1.ExecuteScalar()!);
+
+        // Clean existing data then re-upsert (simulates re-indexing)
+        // 既存データを掃除してから再upsert（再インデックスをシミュレート）
+        _writer.CleanExistingFileData("src/orphan.py");
+        var newId = _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/orphan.py", Lang = "python", Size = 60, Lines = 6,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc).AddMinutes(1),
+        });
+        _writer.InsertChunks([new() { FileId = newId, ChunkIndex = 0, StartLine = 1, EndLine = 6, Content = "def world_replacement(): pass" }]);
+
+        // Old FTS entry should be gone, new one should exist
+        // 旧FTSエントリは消え、新エントリが存在するはず
+        using var cmd2 = _db.Connection.CreateCommand();
+        cmd2.CommandText = "SELECT COUNT(*) FROM fts_chunks WHERE fts_chunks MATCH 'hello_orphan_test'";
+        Assert.Equal(0L, (long)cmd2.ExecuteScalar()!);
+
+        using var cmd3 = _db.Connection.CreateCommand();
+        cmd3.CommandText = "SELECT COUNT(*) FROM fts_chunks WHERE fts_chunks MATCH 'world_replacement'";
+        Assert.Equal(1L, (long)cmd3.ExecuteScalar()!);
+    }
+
+    [Fact]
+    public void PurgeStaleFiles_RemovesDeletedFiles()
+    {
+        // Simulate branch switch: insert a file, then purge when file doesn't exist
+        // ブランチ切り替えをシミュレート: ファイルを挿入後、存在しないファイルをパージ
+        var tempDir = Path.Combine(Path.GetTempPath(), $"codeindex_purge_{Guid.NewGuid():N}");
+        Directory.CreateDirectory(tempDir);
+
+        try
+        {
+            // Create a real file and a "ghost" file entry
+            // 実在するファイルと「ゴースト」ファイルエントリを作成
+            var realFile = Path.Combine(tempDir, "real.py");
+            File.WriteAllText(realFile, "x = 1");
+
+            _writer.UpsertFile(new FileRecord
+            {
+                Path = "real.py", Lang = "python", Size = 5, Lines = 1,
+                Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+            });
+            _writer.UpsertFile(new FileRecord
+            {
+                Path = "ghost.py", Lang = "python", Size = 10, Lines = 2,
+                Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+            });
+
+            var (beforeCount, _, _) = _writer.GetCounts();
+            Assert.Equal(2, beforeCount);
+
+            var purged = _writer.PurgeStaleFiles(tempDir);
+            Assert.Equal(1, purged);
+
+            var (afterCount, _, _) = _writer.GetCounts();
+            Assert.Equal(1, afterCount);
+        }
+        finally
+        {
+            Directory.Delete(tempDir, true);
+        }
+    }
+
+    [Fact]
+    public void DropAll_RemovesAllTables()
+    {
+        // Insert some data, then drop all
+        // データを挿入してから全削除
+        _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/x.py", Lang = "python", Size = 10, Lines = 1,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+        });
+
+        _db.DropAll();
+        _db.InitializeSchema();
+
+        var (files, chunks, symbols) = _writer.GetCounts();
+        Assert.Equal(0, files);
+        Assert.Equal(0, chunks);
+        Assert.Equal(0, symbols);
+    }
+
+    [Fact]
+    public void DeleteFileByPath_RemovesFileAndData()
+    {
+        // Insert a file with chunks and symbols, then delete by path
+        // ファイルとチャンク・シンボルを挿入し、パスで削除
+        var fileId = _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/remove_me.py", Lang = "python", Size = 50, Lines = 5,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+        });
+
+        _writer.InsertChunks([new() { FileId = fileId, ChunkIndex = 0, StartLine = 1, EndLine = 5, Content = "def foo(): pass" }]);
+        _writer.InsertSymbols([new() { FileId = fileId, Kind = "function", Name = "foo", Line = 1 }]);
+
+        var result = _writer.DeleteFileByPath("src/remove_me.py");
+        Assert.True(result);
+
+        var (files, chunks, symbols) = _writer.GetCounts();
+        Assert.Equal(0, files);
+        Assert.Equal(0, chunks);
+        Assert.Equal(0, symbols);
+    }
+
+    [Fact]
+    public void DeleteFileByPath_ReturnsFalseIfNotFound()
+    {
+        // Deleting a non-existent path returns false
+        // 存在しないパスの削除はfalseを返す
+        var result = _writer.DeleteFileByPath("nonexistent/file.py");
+        Assert.False(result);
+    }
+
+    [Fact]
+    public void DeleteFileByPath_DoesNotAffectOtherFiles()
+    {
+        // Deleting one file should not affect another
+        // 1ファイルの削除は他のファイルに影響しない
+        _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/keep.py", Lang = "python", Size = 50, Lines = 5,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+        });
+        _writer.UpsertFile(new FileRecord
+        {
+            Path = "src/delete.py", Lang = "python", Size = 30, Lines = 3,
+            Modified = new DateTime(2025, 6, 1, 0, 0, 0, DateTimeKind.Utc),
+        });
+
+        _writer.DeleteFileByPath("src/delete.py");
+
+        var (files, _, _) = _writer.GetCounts();
+        Assert.Equal(1, files);
+    }
+
+    public void Dispose()
+    {
+        _db.Dispose();
+
+        // Clear SQLite connection pool to release file locks on Windows
+        // Windows環境でファイルロックを解放するためコネクションプールをクリア
+        Microsoft.Data.Sqlite.SqliteConnection.ClearAllPools();
+
+        if (File.Exists(_dbPath))
+            File.Delete(_dbPath);
+    }
+}
