+++ tests/CodeIndex.Tests/FileIndexerTests.cs
@@ -1,0 +1,186 @@
+using CodeIndex.Indexer;
+
+namespace CodeIndex.Tests;
+
+/// <summary>
+/// Tests for FileIndexer.
+/// FileIndexerのテスト。
+/// </summary>
+public class FileIndexerTests
+{
+    [Theory]
+    [InlineData("test.py", "python")]
+    [InlineData("app.js", "javascript")]
+    [InlineData("main.ts", "typescript")]
+    [InlineData("lib.go", "go")]
+    [InlineData("mod.rs", "rust")]
+    [InlineData("App.java", "java")]
+    [InlineData("Service.cs", "csharp")]
+    [InlineData("style.css", "css")]
+    [InlineData("style.scss", "css")]
+    [InlineData("page.vue", "vue")]
+    [InlineData("page.svelte", "svelte")]
+    [InlineData("main.tf", "terraform")]
+    public void DetectLanguage_KnownExtensions_ReturnsCorrectLang(string filename, string expected)
+    {
+        Assert.Equal(expected, FileIndexer.DetectLanguage(filename));
+    }
+
+    [Theory]
+    [InlineData("image.png")]
+    [InlineData("data.bin")]
+    [InlineData("archive.zip")]
+    public void DetectLanguage_UnknownExtensions_ReturnsNull(string filename)
+    {
+        Assert.Null(FileIndexer.DetectLanguage(filename));
+    }
+
+    [Fact]
+    public void ScanFiles_SkipsExcludedDirectories()
+    {
+        // Create a temp directory structure to test scanning
+        // テスト用の一時ディレクトリ構造を作成
+        var tempDir = Path.Combine(Path.GetTempPath(), $"codeindex_test_{Guid.NewGuid():N}");
+        try
+        {
+            Directory.CreateDirectory(tempDir);
+            File.WriteAllText(Path.Combine(tempDir, "app.py"), "print('hello')");
+
+            var nodeModules = Path.Combine(tempDir, "node_modules");
+            Directory.CreateDirectory(nodeModules);
+            File.WriteAllText(Path.Combine(nodeModules, "dep.js"), "module.exports = {}");
+
+            var indexer = new FileIndexer(tempDir);
+            var files = indexer.ScanFiles();
+
+            Assert.Single(files);
+            Assert.Contains("app.py", files[0]);
+        }
+        finally
+        {
+            Directory.Delete(tempDir, true);
+        }
+    }
+
+    [Fact]
+    public void ScanFiles_SkipsExcludedFiles()
+    {
+        var tempDir = Path.Combine(Path.GetTempPath(), $"codeindex_test_{Guid.NewGuid():N}");
+        try
+        {
+            Directory.CreateDirectory(tempDir);
+            File.WriteAllText(Path.Combine(tempDir, "app.js"), "console.log('hello')");
+            File.WriteAllText(Path.Combine(tempDir, "package-lock.json"), "{}");
+
+            var indexer = new FileIndexer(tempDir);
+            var files = indexer.ScanFiles();
+
+            // Only app.js should be found, not package-lock.json
+            // app.jsのみ検出され、package-lock.jsonは除外される
+            Assert.Single(files);
+            Assert.Contains("app.js", files[0]);
+        }
+        finally
+        {
+            Directory.Delete(tempDir, true);
+        }
+    }
+
+    [Fact]
+    public void BuildRecord_CreatesCorrectRecord()
+    {
+        var tempDir = Path.Combine(Path.GetTempPath(), $"codeindex_test_{Guid.NewGuid():N}");
+        try
+        {
+            Directory.CreateDirectory(tempDir);
+            var filePath = Path.Combine(tempDir, "main.py");
+            File.WriteAllText(filePath, "def main():\n    print('hello')\n");
+
+            var indexer = new FileIndexer(tempDir);
+            var (record, content) = indexer.BuildRecord(filePath);
+
+            Assert.Equal("main.py", record.Path);
+            Assert.Equal("python", record.Lang);
+            Assert.Equal(2, record.Lines); // "def main():\n    print('hello')\n" = 2 lines (trailing newline ignored)
+            Assert.NotNull(record.Checksum);
+        }
+        finally
+        {
+            Directory.Delete(tempDir, true);
+        }
+    }
+
+    [Fact]
+    public void ScanFiles_SkipsCaseInsensitiveDirectories()
+    {
+        // SkipDirs should be case-insensitive (e.g. "Build" matches "build")
+        // SkipDirsは大文字小文字を区別しない（例: "Build"は"build"にマッチ）
+        var tempDir = Path.Combine(Path.GetTempPath(), $"codeindex_test_{Guid.NewGuid():N}");
+        try
+        {
+            Directory.CreateDirectory(tempDir);
+            File.WriteAllText(Path.Combine(tempDir, "app.py"), "print('hello')");
+
+            var buildDir = Path.Combine(tempDir, "Build");
+            Directory.CreateDirectory(buildDir);
+            File.WriteAllText(Path.Combine(buildDir, "output.js"), "var x = 1;");
+
+            var indexer = new FileIndexer(tempDir);
+            var files = indexer.ScanFiles();
+
+            Assert.Single(files);
+            Assert.Contains("app.py", files[0]);
+        }
+        finally
+        {
+            Directory.Delete(tempDir, true);
+        }
+    }
+
+    [Fact]
+    public void BuildRecord_CrlfNormalizedToLf()
+    {
+        // CRLF line endings in files should be normalized to LF
+        // ファイル内のCRLF改行はLFに正規化される
+        var tempDir = Path.Combine(Path.GetTempPath(), $"codeindex_test_{Guid.NewGuid():N}");
+        try
+        {
+            Directory.CreateDirectory(tempDir);
+            var filePath = Path.Combine(tempDir, "crlf.py");
+            File.WriteAllBytes(filePath, System.Text.Encoding.UTF8.GetBytes("line1\r\nline2\r\nline3\r\n"));
+
+            var indexer = new FileIndexer(tempDir);
+            var (record, content) = indexer.BuildRecord(filePath);
+
+            Assert.DoesNotContain("\r", content);
+            Assert.Equal(3, record.Lines);
+        }
+        finally
+        {
+            Directory.Delete(tempDir, true);
+        }
+    }
+
+    [Fact]
+    public void BuildRecord_ThrowsForOversizedFile()
+    {
+        // Files exceeding 10 MB should throw InvalidOperationException
+        // 10MBを超えるファイルはInvalidOperationExceptionを投げる
+        var tempDir = Path.Combine(Path.GetTempPath(), $"codeindex_test_{Guid.NewGuid():N}");
+        try
+        {
+            Directory.CreateDirectory(tempDir);
+            var filePath = Path.Combine(tempDir, "large.py");
+            // Create a file just over 10 MB / 10MBを少し超えるファイルを作成
+            var data = new byte[10 * 1024 * 1024 + 1];
+            File.WriteAllBytes(filePath, data);
+
+            var indexer = new FileIndexer(tempDir);
+            Assert.Throws<InvalidOperationException>(() => indexer.BuildRecord(filePath));
+        }
+        finally
+        {
+            Directory.Delete(tempDir, true);
+        }
+    }
+}
